Consolidate services into one
Thinking more about this we want to have all nodes
accept votes and coordinate with eachother on the blocks.
Having these as separate, for now, doesn't make sense.

Maybe in the future, we can have separate services for accepting
votes and for growing the chain. Perhaps there will be more to accepting
votes than to just growing the chain. Or we want to allow anyone to
grow the chain, but not everyone to accept votes into the system.